# framework/deploy/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Builder Framework API
  version: 1.0.0
  description: |
    External REST API for the Knowledge Builder Framework.

    Two interaction models:

    **Consumption Flow** — `POST /api/v1/ask`
    Single endpoint for all knowledge queries. Routes internally through a four-tier
    confidence system (workflow skill → KB retrieval → multi-persona fanout → no-answer).
    Clients never specify which KB, retriever, or persona skill to use.

    **Knowledge Builder Flow** — `/api/v1/kb/authorSkill`
    Server-side stateful session API for authoring new skills. The client is a
    pass-through: it posts user input, shows the server's message, collects the next
    reply, and repeats until `done=true`. The server drives a 14-state machine.

    **Naming conventions:**
    - External API surface (this spec): camelCase field names.
    - DB columns: snake_case (PostgreSQL/Oracle convention). The API layer translates.
    - Python internal code: snake_case (Python convention). Not visible here.

    **Authentication:** Bearer token required on all endpoints.
    Token scopes: `read` (ask + GET authorSkill), `write` (POST/DELETE authorSkill), `admin`.

  contact:
    name: Knowledge Builder Framework Team
  license:
    name: Internal
servers:
  - url: https://{ociVmHostname}
    description: OCI Compute VM (SSE/HTTPS mode — typical for team members)
    variables:
      ociVmHostname:
        default: kb-framework.internal
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/kb/authorSkill:
    post:
      operationId: startAuthorSkillSession
      tags: [AuthorSkill]
      summary: Start a new authoring session
      description: |
        Starts a new knowledge builder session. The server creates a new `synthId`
        and advances the state machine to IDENTIFY_PERSONA.

        If a `synthId` is provided in the request body, the server resumes that session
        instead of creating a new one (equivalent to `POST /api/v1/kb/authorSkill/{synthId}`).

        **Client pattern (pass-through):**
        1. POST user input → receive `{ synthId, state, message, options, ... }`
        2. Show `message` to user
        3. Collect user reply
        4. POST reply to `POST /api/v1/kb/authorSkill/{synthId}`
        5. Repeat until `done=true`

        MCP tool name: `authorSkill` (with `synthId=null`)

        Required scope: `write`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorSkillStartRequest'
            examples:
              newSession:
                summary: Start a new session
                value:
                  input: "I want to automate weekly ADB-S migration status updates for ops engineering"
              resumeSession:
                summary: Resume existing session (alternative to POST /{synthId})
                value:
                  synthId: "synth-ops_eng-adb_s_migration-20260510"
                  input: "resume"
      responses:
        '200':
          description: Session created (or resumed). Returns first state prompt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorSkillResponse'
              examples:
                identifyPersonaState:
                  summary: New session — IDENTIFY_PERSONA state
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    state: "IDENTIFY_PERSONA"
                    message: "I'll help you build that skill. Based on your intent, this looks like an ops_eng skill.\n\nAvailable personas:\n  - ops_eng (Operations Engineer) — 2 skills\n  - tpm (Technical Program Manager) — 3 skills\n\nShall I proceed with ops_eng?"
                    data:
                      suggestedPersona: "ops_eng"
                      suggestionReason: "Intent mentions ADB-S migration, which matches ops_eng skills."
                      availablePersonas:
                        - name: "ops_eng"
                          displayName: "Operations Engineer"
                          skillCount: 2
                        - name: "tpm"
                          displayName: "Technical Program Manager"
                          skillCount: 3
                    options: ["yes / ops_eng", "tpm", "pm", "architect", "list"]
                    progress:
                      step: 1
                      total: 14
                      label: "Identify persona"
                    done: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/UpstreamUnavailable'

    get:
      operationId: listAuthorSkillSessions
      tags: [AuthorSkill]
      summary: List all authoring sessions for the authenticated user
      description: |
        Returns all sessions belonging to the authenticated user, ordered by
        `updatedAt` descending. Includes sessions in all statuses.

        Required scope: `read`
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
              examples:
                mixedSessions:
                  summary: In-progress and completed sessions
                  value:
                    sessions:
                      - synthId: "synth-ops_eng-adb_s_migration-20260510"
                        persona: "ops_eng"
                        skillName: "adb_s_provisioning_migration_status"
                        intent: "Automate weekly ADB-S migration status updates"
                        state: "REVIEW_SCHEMA"
                        progress:
                          step: 5
                          total: 14
                          label: "Review extraction schema"
                        createdAt: "2026-05-10T10:00:00Z"
                        updatedAt: "2026-05-10T14:32:00Z"
                        expiresAt: "2026-05-17T14:32:00Z"
                        status: "in_progress"
                      - synthId: "synth-tpm-weekly_exec-20260509"
                        persona: "tpm"
                        skillName: "weekly_exec_review"
                        intent: "Produce weekly exec review PPT"
                        state: "DONE"
                        progress:
                          step: 14
                          total: 14
                          label: "Done"
                        createdAt: "2026-05-09T09:00:00Z"
                        updatedAt: "2026-05-09T16:00:00Z"
                        expiresAt: null
                        status: "promoted"
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/kb/authorSkill/{synthId}:
    parameters:
      - name: synthId
        in: path
        required: true
        description: Unique session identifier returned by the start endpoint
        schema:
          type: string
          example: "synth-ops_eng-adb_s_migration-20260510"

    post:
      operationId: continueAuthorSkillSession
      tags: [AuthorSkill]
      summary: Continue an in-progress authoring session
      description: |
        Advances the server-side state machine by one step. The client posts the
        user's reply and receives the next prompt.

        The server validates that the input matches the `options` for the current state.
        Invalid input returns HTTP 400 `invalid_input`.

        MCP tool name: `authorSkill` (with `synthId` set)

        Required scope: `write`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorSkillContinueRequest'
            examples:
              confirmPersona:
                summary: Confirm suggested persona
                value:
                  input: "yes"
              provideArtifact:
                summary: Provide artifact path
                value:
                  input: "path/to/example.pptx"
              editFieldDescription:
                summary: Edit extraction instruction for a field
                value:
                  input: "describe migration_status as 'Current migration state for the ADB-S project. One of: canary_running, canary_complete, paused, complete.'"
      responses:
        '200':
          description: Next state prompt after advancing the state machine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorSkillResponse'
              examples:
                analyzeArtifactState:
                  summary: Persona confirmed — ANALYZE_ARTIFACT state
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    state: "ANALYZE_ARTIFACT"
                    message: "Great — ops_eng it is. Do you have an example artifact I can analyze? Provide a file path or URL."
                    data:
                      persona: "ops_eng"
                      acceptedFormats: ["pptx", "docx", "md", "pdf", "confluence_url", "email"]
                    options: ["[file_path_or_url]", "skip — I'll describe the fields manually"]
                    progress:
                      step: 2
                      total: 14
                      label: "Analyze artifact"
                    done: false
                committedState:
                  summary: All artifacts committed — COMMITTED state
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    state: "COMMITTED"
                    message: "Committed. 4 artifacts written to git (sha: a3f1b2c). Want me to validate, ingest, and run eval now?"
                    data:
                      committedPaths:
                        - "framework/parsers/schemas/ops_eng/adb_s_migration_status/v1.json"
                        - "framework/workflow_skills/ops_eng/adb_s_migration_status.yaml"
                      gitCommitSha: "a3f1b2c"
                    options: ["yes / continue", "stop"]
                    progress:
                      step: 9
                      total: 14
                      label: "Committed"
                    done: false
                doneState:
                  summary: Terminal state — skill promoted to production
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    state: "DONE"
                    message: "Done. ops_eng.adb_s_migration_status is now in production. It will run every Monday at 9am and is available on request via askKnowledgeBase."
                    data: {}
                    options: []
                    progress:
                      step: 14
                      total: 14
                      label: "Done"
                    done: true
        '400':
          description: |
            Invalid input — does not match any accepted option for the current state.
            Also returned for invalid state transitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "invalid_input"
                  message: "Input 'banana' is not a valid option for state IDENTIFY_PERSONA. Valid options: yes, ops_eng, tpm, pm, architect, list"
                  details: {}
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          description: Session not found or belongs to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "session_not_found"
                  message: "Session synth-ops_eng-adb_s_migration-20260510 not found or you do not have access."
                  details: {}
        '409':
          description: Session is expired or already done
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                sessionExpired:
                  value:
                    error:
                      code: "session_expired"
                      message: "Session TTL has elapsed. Start a new session."
                      details: {}
                sessionDone:
                  value:
                    error:
                      code: "session_done"
                      message: "Session is in DONE state. No further input accepted."
                      details: {}
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/UpstreamUnavailable'

    get:
      operationId: getAuthorSkillSession
      tags: [AuthorSkill]
      summary: Get the current state of a session (without advancing it)
      description: |
        Returns the current state and data of the session. Used to inspect a session
        before resuming it — for example, after the user closes and reopens Claude Code.

        Returns the same envelope as POST responses, with an additional `sessionSummary`
        field showing accumulated session context.

        Required scope: `read`
      responses:
        '200':
          description: Current session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorSkillGetResponse'
              examples:
                reviewSchemaState:
                  summary: Session at REVIEW_SCHEMA state
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    state: "REVIEW_SCHEMA"
                    message: "Review the extraction instructions for your 3 fields. Edit any that are unclear. Type 'ok' to confirm all."
                    data:
                      fields:
                        - name: "migration_status"
                          type: "string"
                          description: "Status of OC1 pod migration from ExaCS to ADB-S."
                          enum: null
                          example: "canary_done_emea_paused_us"
                        - name: "pod_count"
                          type: "integer"
                          description: "Total number of pods migrated to ADB-S in this reporting period."
                          enum: null
                          example: 47
                      schemaNotes: "The 'description' field is what the LLM parser uses as its extraction instruction."
                    options: ["ok", "describe <field> as <text>", "set type of <field> to <type>", "show <field>"]
                    progress:
                      step: 5
                      total: 14
                      label: "Review extraction schema"
                    done: false
                    sessionSummary:
                      persona: "ops_eng"
                      intent: "Automate weekly ADB-S migration status updates"
                      artifactPath: "path/to/example.pptx"
                      fieldsConfirmed: ["migration_status", "pod_count", "canary_region"]
                      sourcesConfigured: []
                      triggersConfigured: null
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          description: Session not found or belongs to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      operationId: cancelAuthorSkillSession
      tags: [AuthorSkill]
      summary: Cancel / abandon an in-progress session
      description: |
        Marks the session as `abandoned`. No git changes are rolled back — if the session
        was in COMMITTED or later state, those commits are kept.

        Sessions in `abandoned` state are cleaned up after 7 days.

        Required scope: `write`
      responses:
        '200':
          description: Session cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelSessionResponse'
              examples:
                cancelledBeforeCommit:
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    status: "abandoned"
                    message: "Session abandoned. No files were committed."
                    committedBeforeAbandon: false
                cancelledAfterCommit:
                  value:
                    synthId: "synth-ops_eng-adb_s_migration-20260510"
                    status: "abandoned"
                    message: "Session abandoned. Previously committed artifacts are retained in git."
                    committedBeforeAbandon: true
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          description: Session not found or belongs to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─────────────────────────────────────────────────────────────────────────
  # Operational endpoints
  # ─────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token issued per consumer. Stored in OCI Vault.
        Format: `Authorization: Bearer <token>`

        Token scopes:
        - `read` — askKnowledgeBase and GET /api/v1/kb/authorSkill* endpoints
        - `write` — authorSkill (POST) and DELETE /api/v1/kb/authorSkill/{synthId}
        - `admin` — metrics and system configuration endpoints
  schemas:
...
    AuthorSkillStartRequest:
      type: object
      required: [input]
      properties:
        input:
          type: string
          minLength: 1
          maxLength: 4096
          description: |
            The user's initial intent message describing the skill to build.
            Example: "I want to automate weekly ADB-S migration status updates for ops engineering"
        synthId:
          type: [string, 'null']
          description: |
            If provided, resumes this session instead of creating a new one.
            Equivalent to calling `POST /api/v1/kb/authorSkill/{synthId}`.
    AuthorSkillContinueRequest:
      type: object
      required: [input]
      properties:
        input:
          type: string
          minLength: 1
          maxLength: 4096
          description: |
            The user's reply to the server's last prompt.
            Must match one of the patterns in the `options` array of the previous response.
            Example: "yes", "ops_eng", "path/to/artifact.pptx", "ok"

    # ── AuthorSkill — responses ──────────────────────────────────────────────
    AuthorSkillResponse:
      type: object
      required: [synthId, state, message, options, progress, done]
      description: |
        Universal session response envelope. Returned by all POST authorSkill endpoints
        regardless of which state the machine is in.
      properties:
        synthId:
          type: string
          description: |
            Unique session identifier. Created on the first POST, returned on every
            subsequent call. The client must store this to resume the session.
          example: "synth-ops_eng-adb_s_migration-20260510"
        state:
          $ref: '#/components/schemas/SessionState'
        message:
          type: string
          description: |
            Plain English text to show the user. No client formatting required.
            This is what Claude Code / Codex displays verbatim.
        data:
          type: object
          description: |
            State-specific payload. Schema varies by state — see PDD V3 §6.4 for
            the per-state data shape.
          additionalProperties: true
        options:
          type: array
          items:
            type: string
          description: |
            Accepted input patterns for this state. Used by Claude Code to guide the user
            on what to type next.
        progress:
          $ref: '#/components/schemas/SessionProgress'
        done:
          type: boolean
          description: |
            True only in the DONE state. When true, the session is complete and archived.
            No further input is accepted.
    AuthorSkillGetResponse:
      allOf:
        - $ref: '#/components/schemas/AuthorSkillResponse'
        - type: object
          properties:
            sessionSummary:
              $ref: '#/components/schemas/SessionSummary'
...
    SessionListResponse:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionListItem'
          description: All sessions for the authenticated user, ordered by updatedAt descending.
...
  responses:
    BadRequest:
      description: |
        Bad request. Common causes:
        - question is empty or exceeds 4096 chars (ask endpoint)
        - input does not match accepted options for current state (authorSkill)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "invalid_argument"
              message: "question must be between 1 and 4096 characters"
              details: {}
    Unauthenticated:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthenticated"
              message: "Authorization header missing or token invalid"
              details: {}
    PermissionDenied:
      description: Token lacks the required scope for this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "permission_denied"
              message: "Token lacks 'write' scope required for this operation"
              details: {}
    RateLimited:
      description: Per-token RPM cap exceeded.
      headers:
        Retry-After:
          description: Seconds until the rate limit window resets.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "rate_limited"
              message: "Rate limit exceeded. Retry after 23 seconds."
              details:
                retryAfterSeconds: 23
    UpstreamUnavailable:
      description: ADB or OCI GenAI is unreachable. Retry with exponential backoff.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "upstream_unavailable"
              message: "Oracle ADB is currently unreachable. The framework is in read-only mode."
              details:
                affectedComponent: "adb"